ioteca_service/requirements.txt
Django==1.9.6
django-braces==1.8.1
django-cors-headers==1.1.0
django-oauth-toolkit==0.10.0
djangorestframework==3.3.3
docutils==0.12
oauthlib==1.0.3
Pillow==3.2.0
six==1.10.0
